Fix bug where killed trx doesn't release concurrency control semaphore
In between a successful `sem_trywait` or `sem_wait` and the lock record
getting created with `lock->fought` set to TRUE, if the query is killed
(e.g. by proxysql `default_query_timeout`), then the semaphore wouldn't
be released. Once this has happened enough times for a page, all the
permits would be gone, resulting in transactions getting stuck forever.

Note that we haven't actually seen the above race condition happens.
However, the previous implementation could trigger the bug easily due
to the blocking nature of `sem_wait`:

1. trx calls `sem_wait` to wait for the concurrency control semaphore.
2. trx exceeds `default_query_timeout`, proxysql executes KILL QUERY.
3. mysql executes `innobase_kill_connection`.
4. (an eternity later) `sem_wait` returns 0, trx proceeds to do another
   round of `row_search_for_mysql` with `fought` set to TRUE.
5. as the thread was already killed, `row_search_for_mysql` returns
   `DB_INTERRUPTED`.

As it is tricky to interrupt `sem_wait` to stop it from acquiring the
semaphore [1], we change the implementation to do `sem_trywait` in a
loop instead. This is of course less efficient compared to the futex
implementation of `sem_wait`. From testing, a sleep time of 10ms in
between `sem_trywait` calls seems to work ok under high concurrency.
This can be controlled via `innodb_concurrency_control_sleep_time_us`,
with the default of 10000 (i.e. 10ms).

[1] https://stackoverflow.com/questions/61448888/why-does-sem-wait-not-unblock-and-return-1-on-an-interrupt