Don't crash the server when an INSERT exceeds concurrency control limit
This code path seems to be triggered for INSERT statements only, which
would call either of the following methods directly:
- `lock_clust_rec_modify_check_and_lock`
- `lock_sec_rec_modify_check_and_lock`

For comparison, UPDATE / DELETE / SELECT ... FOR UPDATE statements
would call `row_search_for_mysql` first, which would then call either
of the following methods first:
- `lock_clust_rec_read_check_and_lock`
- `lock_sec_rec_read_check_and_lock`

As all the logics are inside `row_search_for_mysql`, e.g. enter global
queue and increment metrics, it means an INSERT statement that fails to
acquire the concurrency control semaphore in `lock_rec_lock_slow` would
error out immediately (i.e. the code path in this commit) and wouldn't
increment any of concurrency control metrics. This should be rare, so
we leave it as-is for now.